requirements.txt
Django==1.8.5
requests==2.9.0
subprocess32==3.2.7
#pyOpenSSL==0.13
#pyasn1==0.1.9
#ndg-httpsclient==0.4.0

# xml, oai
beautifulsoup4==4.4.1
lxml==3.4.4
Sickle==0.5

# database
psycopg2==2.6.1

# rest
Markdown==2.6.5
djangorestframework==3.3.2
